+++ CHANGELOG.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+10.3.0, 2025-04-30
+------------------
+- In ``cleaners.drop_zombies``, additionally dropped agencies without routes.
+- Fixed ``routes.get_routes`` to handle the case when at least one of a route's shapes is a point.
+- Turned the example notebook into a Marimo notebook.
+
+10.2.2, 2025-03-13
+------------------
+- Bugfixed an edge case in the function ``shapes.split_simple``.
+
 10.2.1, 2025-03-12
 ------------------
 - Added a forgotten ``drop_duplicates('shape_id')`` line to the function ``miscellany.compute_screen_line_counts``, which speeds things up on large sets of screen lines.